requirements.txt
pandas
openpyxl
python-barcode
Pillow